Fix priors back as they were
The risk assortativity adjustment did not
work as intended because package='syphLAMA' was written
into the priors <<- read.delim(...) command.

Since we found out that partnerships are actually a
larger cause of why Louisiana reacts differently to
MSM targeted interventions than assortativity parameters,
I am thinking as this time we will not try to
adjust these parameters as we were planning to.