Overwrite natural_history_parameters earlier
In this version of get_best_simultaneous_optimizers, we produce
a data frame of 204 columns (2*length(theta)) for the top 25
best optimizers for the simultaneous model.

Before saving that top25_best_simultaneous_params.rds file,
the natural history parameters are overwritten (to match
how the simultaneous dLogPosterior function works (it uses the
second set of natural history parameters for both states)).

This commit removes the older code which overwrites the natural
history parameters *after* finding the top25 best simultaneous
parameters.