Adding code to visualize interventions outcomes
This code is an improved version of the code that's in the
interventions Rmarkdown in the mini-manuscript.

I should re-do the mini-manuscript Rmarkdown documents using these
functions soon.

Note: It would also be good to set up save=TRUE or FALSE as a function
argument that can be passed to some of these simulate_and_plot*
functions.

Note 2: I think it would be better to have a more comprehensive
extract_* function for extracting statistics from the intervention
scenarios, so hopefully that's soon to be added and that will clean
up how there's so many different versions of simulate_and_plot* code.