Fix issues in plotting posteriors:
dunif was used where dnorm was needed in the duration of secondary
syphilis.

the bezier.fun was defaulted to outputting weekly because yearly
discontinuities in screening were causing (for a lack of better words)
bouncy trends in prevalence/incidence. the bezier.fun was changed to
have a length.out parameter which is used in plot.posteriors for
yearly outputs to keep plotting consistent with prior formatting.