Adjust plot_posteriors update to behav.lin
This is the second part of the update for the behav.lin param;
Previously the priors.txt and syph_trans_model.cpp were updated so that
c_rr (the time-serries version of behav.lin) is a fraction of (1-b[2])
(the remaining possible transmission probability, where b[2] is the
fixed transmission probability) that is tacked onto transmission during
simulations.

This setup i.e. transmission = (b[2] + (1-b[2])*c_rr) ensures that
transmission probability never go above 1 and are in fact biased towards
no growth because 0 is the mode and median value of the new prior,
beta(1,15).